requirements.txt
fastapi
uvicorn
python-multipart
requests
python-dotenv
openai
gradium
jinja2
websockets
